% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_atoms.R
\docType{data}
\name{db_atoms}
\alias{db_atoms}
\title{OES atoms spectral lines from NIST database}
\format{A data frame with xxxxxxx rows and yy variables:
\describe{
\item{X1}{}
\item{element}{}
\item{wavelength}{}
\item{rel.}{}
\item{aki}{}
\item{fik}{}
\item{s}{}
\item{log_gf}{}
\item{acc}{}
\item{ei}{}
\item{ek}{}
\item{ji}{}
\item{jk}{}
\item{gi}{}
\item{gk}{}
\item{configurations}{}
\item{terms}{}
\item{type}{}
\item{tp_refs}{}
\item{line_refs}{}
}}
\source{
\url{http://nist/}
}
\usage{
db_atoms
}
\description{
A dataset containing the OES spectral lines data
}
\keyword{datasets}
