% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_molecules.R
\docType{data}
\name{db_molecules}
\alias{db_molecules}
\title{OES molecules spectral lines from NIST database}
\format{A data frame with xxxxxxx rows and yy variables:
\describe{
\item{X1}{}
\item{wavelength}{wavelength}
\item{frequency}{}
\item{intensity}{}
\item{vc}{}
\item{vcc}{}
\item{s}{}
\item{ec}{}
\item{note}{}
...
}}
\source{
\url{http://nist/}
}
\usage{
db_molecules
}
\description{
A dataset containing the OES spectral lines data
}
\keyword{datasets}
