% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_nist.R
\name{nist_db_content}
\alias{nist_db_content}
\title{list content of NIST OES database}
\usage{
nist_db_content()
}
\value{
list of files in NIST database folder
}
\description{
list content of NIST OES database
}
\examples{
\dontrun{nist_db_content()}
}
\author{
K. Juraic
}
