% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{oes_lst_2_mat}
\alias{oes_lst_2_mat}
\title{Convert OES list to matrix}
\usage{
oes_lst_2_mat(oes_lst, plot = TRUE)
}
\arguments{
\item{oes_lst}{list of oes data}

\item{plot}{TRUE plot matrix}
}
\description{
Convert OES list to matrix
}
\examples{
       \dontrun{oes_lst_2_mat(oes_lst)}
}
\author{
K. Juraic
}
