% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{oes_lst_2_time}
\alias{oes_lst_2_time}
\title{OES time scale from list}
\usage{
oes_lst_2_time(oes_lst)
}
\arguments{
\item{oes_lst}{List with OES data}
}
\value{
OES time scale in array
}
\description{
Extract time scale from OES data list
}
\examples{
     \dontrun{oes_lst_2_time(oes_lst)}
}
\author{
K. Juraic
}
