% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oes_analysis.R
\name{oes_plot_compare}
\alias{oes_plot_compare}
\title{Compare experimental OES data to NIST literature}
\usage{
oes_plot_compare(nist, oes_lst)
}
\arguments{
\item{nist}{data.frame with NIST data(imported with read_nist_oes_multi)}

\item{oes_lst}{list with experimental data (imported with read_file_2_list())}
}
\value{
NA
}
\description{
Compare experimental OES data to NIST literature
}
\examples{
\dontrun{oes_plot_compare(nist, oes_lst)}
}
\author{
K. Juraic
}
