% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read_file_2_list}
\alias{read_file_2_list}
\title{Read data from multiple files to list}
\usage{
read_file_2_list(
  oes_folder = tcltk::tk_choose.dir(caption = "Select OES folder:"),
  read_function = read_ocean_tab,
  all_files = FALSE,
  file_pattern = "*.*"
)
}
\arguments{
\item{oes_folder}{Folder path from where to read}

\item{read_function}{Functio tpo be used for single file read}

\item{all_files}{TRUE all files satisfieing filePattern in folder will be read}

\item{file_pattern}{File pattern}
}
\value{
list of read data
}
\description{
Read data from multiple files to list by using specified function.
             If the folder name is not specified GUI dialog will be opened to
             select folder path. Can be read all files in folder or selected by
             dialog.
}
\examples{
\dontrun{dat = read_file_2_list()}
}
\author{
K. Juraic
}
