% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_nist.R
\name{read_nist_oes}
\alias{read_nist_oes}
\title{Read optical spectral lines data from file for single element}
\usage{
read_nist_oes(file_name)
}
\arguments{
\item{file_name}{spectral data file name}
}
\value{
data.frame with optical spectral line data
}
\description{
Read optical spectral lines data from file for single element
}
\author{
K. Juraic
}
