% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_nist.R
\name{read_nist_oes_multi}
\alias{read_nist_oes_multi}
\title{read multiple oes file}
\usage{
read_nist_oes_multi(element_list, db_path = "extdata/Database/")
}
\arguments{
\item{element_list}{list of elements to read data}

\item{db_path}{path to directory with oes files}
}
\value{
data.frame with spectral data
}
\description{
read multiple oes file
}
\examples{
      \dontrun{read_nist_oes_multi(c("AR_I","O_I"))}
}
\author{
K. Juraic
}
