% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read_ocean_tab}
\alias{read_ocean_tab}
\title{Read Ocean optics spectra from file}
\usage{
read_ocean_tab(file_name)
}
\arguments{
\item{file_name}{File name for stored ocean optics spectra}
}
\value{
list(name, data, scan_time, int_time, spectra_avg, boxcar, n_pixels)
}
\description{
Read ocean optics spectra from tab delimited file with header
}
\examples{
    \dontrun{read_ocean_tab("ocean_file.txt")}
}
\author{
K. Juraic
}
