% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roes.R
\docType{package}
\name{roes}
\alias{roes}
\title{roes: A package for optical emission spectroscopy (OES) data analysis.}
\description{
The roes package provides three categories of important functions:
for import experimental data,
for preprocesing OES data and
for comparison with NIST database.
}
\section{Import}{

Import ...
}

\section{Preprocessing}{

Preprocessing ...
}

\section{Analysis}{

Analysis ...
}

