% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read_ocean_tab}
\alias{read_ocean_tab}
\title{Read Ocean optics spectra from file}
\usage{
read_ocean_tab(fileName)
}
\arguments{
\item{fileName}{File name for stored ocean optics spectra}
}
\value{
list(name, data, scanTime, intTime, spectraAvg, boxcar, nPixels)
}
\description{
Read ocean optics spectra from tab delimited file with header
}
\examples{
    \dontrun{read_ocean_tab("ocean_file.txt")}
}
\author{
K. Juraic
}
